+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 
 ### Fixed
 - Issue #1385: Add missing internationalization calls for layer names.
+- Issue #1413: Fix legend toggle for featureless vector layers.
 - Issue #1422: Reset layers in CompareMaps on any deselection to avoid bugs regarding which layers are visible on what side of the comparison.
 - Documentation: corrected usage of 'treeType' (version 2).